Add files via upload
I recently completed a fascinating project on DataCamp titled "What and Where are the World's Oldest Businesses." In this eye-opening course, I delved into the intriguing history of businesses that have stood the test of time. Through engaging lessons and hands-on exercises, I learned how to analyze historical data and discover the oldest businesses worldwide. It was a thrilling journey through centuries of entrepreneurship, uncovering the secrets behind their longevity. This project not only expanded my data analysis skills but also deepened my appreciation for the enduring spirit of businesses across the globe.